Move FromVariant macro impl to the separate file
